--- pkg/deployments/tasks/20230206-aave-rebalanced-linear-pool-v4/input.ts
@@ -1,31 +1,0 @@
-import Task, { TaskMode } from '../../src/task';
-import { MONTH } from '@balancer-labs/v2-helpers/src/time';
-
-export type AaveLinearPoolDeployment = {
-  Vault: string;
-  BalancerQueries: string;
-  ProtocolFeePercentagesProvider: string;
-  WETH: string;
-  FactoryVersion: string;
-  PoolVersion: string;
-  InitialPauseWindowDuration: number;
-  BufferPeriodDuration: number;
-};
-
-const Vault = new Task('20210418-vault', TaskMode.READ_ONLY);
-const BalancerQueries = new Task('20220721-balancer-queries', TaskMode.READ_ONLY);
-const ProtocolFeePercentagesProvider = new Task('20220725-protocol-fee-percentages-provider', TaskMode.READ_ONLY);
-const WETH = new Task('00000000-tokens', TaskMode.READ_ONLY);
-
-const BaseVersion = { version: 4, deployment: '20230206-aave-rebalanced-linear-pool-v4' };
-
-export default {
-  Vault,
-  BalancerQueries,
-  ProtocolFeePercentagesProvider,
-  WETH,
-  FactoryVersion: JSON.stringify({ name: 'AaveLinearPoolFactory', ...BaseVersion }),
-  PoolVersion: JSON.stringify({ name: 'AaveLinearPool', ...BaseVersion }),
-  InitialPauseWindowDuration: MONTH * 3,
-  BufferPeriodDuration: MONTH,
-};
