--- pkg/deployments/tasks/20230206-aave-rebalanced-linear-pool-v4/output/gnosis.json
@@ -1,6 +1,0 @@
-{
-  "AaveLinearPoolFactory": "0x9dA18982a33FD0c7051B19F0d7C76F2d5E7e017c",
-  "MockAaveLendingPool": "0x81cFAE226343B24BA12EC6521Db2C79E7aeeb310",
-  "MockStaticAToken": "0xa523f47A933D5020b23629dDf689695AA94612Dc",
-  "MockAaveLinearPool": "0x225E0047671939A8d78e08EBd692788Abe63f15c"
-}
