--- pkg/deployments/tasks/20230206-aave-rebalanced-linear-pool-v4/readme.md
@@ -1,16 +1,0 @@
-# 2023-02-06 - Aave Rebalanced Linear Pool v4
-
-Deployment of the `AaveLinearPoolFactory`, for Linear Pools with a wrapped aToken.
-
-Supersedes `20221207-aave-rebalanced-linear-pool-v3`, fixing the reentrancy issue described in this [forum post](https://forum.balancer.fi/t/reentrancy-vulnerability-scope-expanded/4345).
-
-## Useful Files
-
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [Polygon mainnet addresses](./output/polygon.json)
-- [Arbitrum mainnet addresses](./output/arbitrum.json)
-- [Optimism mainnet addresses](./output/optimism.json)
-- [BSC mainnet addresses](./output/bsc.json)
-- [Gnosis mainnet addresses](./output/gnosis.json)
-- [Goerli testnet addresses](./output/goerli.json)
-- [`AaveLinearPoolFactory` artifact](./artifact/AaveLinearPoolFactory.json)
