+++ pkg/deployments/tasks/20230410-aave-linear-pool-v5/readme.md
@@ -1,0 +1,11 @@
+# 2023-04-10 - Aave Linear Pool v5
+
+Deployment of the `AaveLinearPoolFactory`, for Linear Pools with a wrapped aToken.
+
+Supersedes `20221207-aave-rebalanced-linear-pool-v3`, fixing the reentrancy issue described in this [forum post](https://forum.balancer.fi/t/reentrancy-vulnerability-scope-expanded/4345).
+
+## Useful Files
+
+- [`AaveLinearPool` artifact](./artifact/AaveLinearPool.json)
+- [`AaveLinearPoolFactory` artifact](./artifact/AaveLinearPoolFactory.json)
+- [`AaveLinearPoolRebalancer` artifact](./artifact/AaveLinearPoolRebalancer.json)
